Features: 	- EM 8 Static IRS Configurator: Day Counts, Frequencies, and Tenors Pre-set for CLP, CZK, HKD, HUF, MXN, PLN, SGD, and ZAR 	- Segmented Intrisic Time Value Explain - Time Roll, Maturity Roll Up, and Maturity Roll down Components 	- Transform of the Log Normal Volatility to Shifted Log Normal Volatility
Bug Fixes:
	- Integrated Cross Volatility Quanto Adjustment Connect-in fix
	- Bug Fix - 1M Maturity Roll down Swap Date and Maturity Roll Down PnL
	- Bug Fix - 3M Maturity Roll down Swap Date and Maturity Roll Down PnL

Samples:
	- FRAMkt: Expand FRAMarketComponent sample valuation to incorporate the FRAStandardComponent metrics
	- FRAMkt Sample: Incorporation of a real market EONIA and EURIBOR6M calibration instruments and curve construction in the FRA Standard/Market Valuation
	- FRAMktVolAnalysis: Extract the FRA shifted lognormal volatility analysis impact on the market forwards and convexity correction
	- FRAMktVolAnalysis: Incorporation of the explicit market log normal ATM Vols for EONIA and EURIBOR laid out in Mercurio (2010)

+++ CreditAnalytics/dev/src/org/drip/analytics/date/DateTime.java
@@ -119,13 +119,13 @@ public class DateTime extends org.drip.service.stream.Serializer {
 			org.drip.service.stream.Serializer.NULL_SER_STRING.equalsIgnoreCase (astrField[1]))
 			throw new java.lang.Exception ("DateTime de-serializer: Cannot locate long time");
 
-		_lTime = new java.lang.Long (astrField[1]);
+		_lTime = new java.lang.Long (astrField[1]).longValue();
 
 		if (null == astrField[2] || astrField[2].isEmpty() ||
 			org.drip.service.stream.Serializer.NULL_SER_STRING.equalsIgnoreCase (astrField[2]))
 			throw new java.lang.Exception ("DateTime de-serializer: Cannot locate date");
 
-		_dblDate = new java.lang.Double (astrField[2]);
+		_dblDate = new java.lang.Double (astrField[2]).doubleValue();
 	}
 
 	/**
